requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==20.8b1
cairocffi===file-.cairocffi-VERSION
CairoSVG==2.5.1
cffi==1.14.5
click==7.1.2
cssselect2==0.4.1
dataclasses==0.8
defusedxml==0.6.0
Django==3.1.1
django-debug-toolbar==2.2
django-environ==0.4.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
flake8==3.8.4
gunicorn==20.0.4
html5lib==1.1
importlib-metadata==3.4.0
mccabe==0.6.1
more-itertools==8.4.0
mypy-extensions==0.4.3
numpy==1.19.5
packaging==20.4
pandas==1.1.5
pathspec==0.8.1
Pillow==7.2.0
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
Pyphen==0.10.0
pytest==5.4.3
pytest-django==3.10.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.11.13
reportlab==3.5.59
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tinycss2==1.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wcwidth==0.2.5
WeasyPrint==52.2
webencodings==0.5.1
zipp==3.4.0
